+++ lib/generator.dart
@@ -1,5 +1,5 @@
 import 'package:build/build.dart';
-import 'package:crimson/src/crimson_generator.dart';
+import 'package:crimson/src/generator/generator.dart';
 import 'package:source_gen/source_gen.dart';
 
 /// A builder which generates Crimson code.